Use folders instead of risible name prefixes
The PDS Engineering Node's fixation with prefixing instead of namespaces makes the file tree widget of Jupyter unusable because all you see is `pds-api-cli…` instead of actual files, forcing you to play with the widget instead of the notebooks. This commit rearranges the notebooks into a useful and browseable file hierarchy, each of which makes its own namespace without the need for inane prefixing.